Remove ContainsNotExistsElementBenchmark from the caliper suite
ContainsNotExistsElementBenchmark does not measure what it is supposed to: every "collisionless" element in the control set actually has at least one same hash code twin in a the tested set.
I can not find a way to implement it right in the case of the managed hash code collisions (same hash code, same hash table index) so I let it go.